feat: refresh tables after drop/alter table
