+++ src/views/dashboard/modules/table-manager.vue
@@ -150,14 +150,21 @@ a-card.table-manager(:bordered="false")
 
   const route = useRoute()
   const { insertNameToPyCode } = usePythonCode()
-  const { tablesSearchKey, tablesTreeData, tablesTreeRef, refreshTables } = useSiderTabs()
+  const {
+    tablesSearchKey,
+    tablesTreeData,
+    tablesTreeRef,
+    isRefreshingDetails,
+    refreshTables,
+    loadMore,
+    loadMoreColumns,
+  } = useSiderTabs()
   const { tablesLoading, originTablesTree } = storeToRefs(useDataBaseStore())
   const { getTableByName, getTables, addChildren, generateTreeChildren } = useDataBaseStore()
 
   const LAYOUT_PADDING = 16
   const HEADER = 58
   const listHeight = LAYOUT_PADDING * 3 + HEADER
-  const isRefreshingDetails = ref<{ [key: number]: boolean }>({ 0: false })
 
   const expandedKeys = ref<number[]>()
 
@@ -179,114 +186,6 @@ a-card.table-manager(:bordered="false")
       }
     })
   })
-
-  const loadMoreColumns = (nodeData: TableTreeParent) =>
-    new Promise<void>((resolve, reject) => {
-      getTableByName(nodeData.title)
-        .then((result: any) => {
-          const { output } = result
-          const {
-            records: {
-              rows,
-              schema: { column_schemas: columnSchemas },
-            },
-          } = output[0]
-          const { treeChildren, timeIndexName } = generateTreeChildren(nodeData, rows, columnSchemas)
-          addChildren(nodeData.key, treeChildren, timeIndexName)
-          resolve()
-        })
-        .catch(() => {
-          reject()
-        })
-    })
-
-  const loadMore = async (nodeData: TableTreeParent) => {
-    if (nodeData.childrenType === 'details') {
-      isRefreshingDetails.value[nodeData.key] = true
-
-      const createTable = new Promise<object>((resolve, reject) => {
-        editorAPI
-          .runSQL(`show create table "${nodeData.title}"`)
-          .then((res: any) => {
-            const sql = `${res.output[0].records.rows[0][1]}`
-            const regex = /ttl = '(\w)+'/g
-            const ttl = sql.match(regex)?.[0].slice(7, -1) || '-'
-            const result = {
-              key: 'createTable',
-              value: { sql, ttl },
-            }
-            resolve(result)
-          })
-          .catch(() => {
-            const result = {
-              key: 'createTable',
-              value: { sql: '-', ttl: '-' },
-            }
-            reject(result)
-          })
-      })
-
-      const rowAndTime = new Promise<object>((resolve, reject) => {
-        const getRowAndTime = () => {
-          editorAPI
-            .runSQL(
-              nodeData.timeIndexName !== '%TIMESTAMP%'
-                ? `select count(*), min (${nodeData.timeIndexName}), max (${nodeData.timeIndexName}) from "${nodeData.title}"`
-                : `select count(*) from "${nodeData.title}"`
-            )
-            .then((res: any) => {
-              const resArray = res.output[0].records.rows[0]
-              const timestampType = res.output[0].records.schema.column_schemas[1]?.data_type || '-'
-              const result = {
-                key: 'rowAndTime',
-                value: { rowCount: resArray[0], min: resArray[1] || '-', max: resArray[2] || '-', timestampType },
-              }
-              resolve(result)
-            })
-            .catch(() => {
-              const result = {
-                key: 'rowAndTime',
-                value: { rowCount: '-', min: '-', max: '-' },
-              }
-              reject(result)
-            })
-        }
-        if (!nodeData.timeIndexName) {
-          loadMoreColumns(nodeData).then(() => getRowAndTime())
-        } else {
-          getRowAndTime()
-        }
-      })
-
-      return Promise.allSettled([rowAndTime, createTable]).then((result: any[]) => {
-        const children: TableDetail[] = []
-        const rowAndTimeResult = result[0].value || result[0].reason
-        const createTableResult = result[1].value || result[1].reason
-        children.push({
-          key: `${nodeData.title}.details.${rowAndTimeResult.key}`,
-          title: rowAndTimeResult.key,
-          parentKey: nodeData.key,
-          tableName: nodeData.title,
-          isLeaf: true,
-          info: { ...rowAndTimeResult.value, ttl: createTableResult.value.ttl },
-          class: 'details',
-        })
-
-        children.push({
-          key: `${nodeData.title}.details.${createTableResult.key}`,
-          title: createTableResult.key,
-          parentKey: nodeData.key,
-          tableName: nodeData.title,
-          isLeaf: true,
-          info: createTableResult.value,
-          class: 'details',
-        })
-        addChildren(nodeData.key, children, nodeData.timeIndexName, 'details')
-        isRefreshingDetails.value[nodeData.key] = false
-      })
-    }
-    return loadMoreColumns(nodeData)
-  }
 
   const INSERT_MAP: { [key: string]: any } = {
     // query: insertNameToQueryCode,
